Move HelpOption to ZendeskHelper.kt
